long-term-storage/requirements.txt
boto3
pandas
pymssql
pylint
pytest
python-dotenv